requirements.txt
psycopg2-binary
langchain
langchain-groq
python-dotenv
groq